+++ core/parse.c
@@ -16,7 +16,7 @@ int metric = 1;
 int diveid = -1;
 
 event_allocation_t event_allocation = { .event.deleted = 1 };
-
+struct parser_settings cur_settings;
 
 struct divecomputer *cur_dc = NULL;
 struct dive *cur_dive = NULL;
@@ -469,7 +469,7 @@ void add_dive_site(char *ds_name, struct dive *dive)
 		}
 		if (ds) {
 			// we have a uuid, let's hope there isn't a different name
-			if (same_string(ds->name, "")) {
+			if (empty_string(ds->name)) {
 				ds->name = copy_string(buffer);
 			} else if (!same_string(ds->name, buffer)) {
 				// if it's not the same name, it's not the same dive site