+++ mobile-widgets/qml/DiveList.qml
@@ -162,9 +162,8 @@ Kirigami.ScrollablePage {
 						anchors.fill: parent
 						enabled: parent.visible
 						onClicked: {
-							parent.visible = false
+							deleteButtonVisible = false
 							timer.stop()
-							detailsWindow.showDiveIndex(index)
 							manager.deleteDive(dive.id)
 						}
 					}
@@ -332,7 +331,9 @@ Kirigami.ScrollablePage {
 	}
 
 	property QtObject downloadFromDCAction: Kirigami.Action {
-		iconName: "downloadDC"
+		icon {
+			name: "downloadDC"
+		}
 		onTriggered: {
 			downloadFromDc.dcImportModel.clearTable()
 			stackView.push(downloadFromDc)
@@ -340,14 +341,18 @@ Kirigami.ScrollablePage {
 	}
 
 	property QtObject addDiveAction: Kirigami.Action {
-		iconName: "list-add"
+		icon {
+			name: "list-add"
+		}
 		onTriggered: {
 			startAddDive()
 		}
 	}
 
 	property QtObject saveAction: Kirigami.Action {
-		iconName: "document-save"
+		icon {
+			name: "document-save"
+		}
 		onTriggered: {
 			Qt.inputMethod.hide()
 			startPage.saveCredentials();
@@ -355,7 +360,9 @@ Kirigami.ScrollablePage {
 	}
 
 	property QtObject offlineAction: Kirigami.Action {
-		iconName: "qrc:/qml/nocloud.svg"
+		icon {
+			name: "qrc:/qml/nocloud.svg"
+		}
 		onTriggered: {
 			manager.syncToCloud = false
 			manager.credentialStatus = QMLManager.CS_NOCLOUD