+++ qt-models/diveplannermodel.cpp
@@ -151,7 +151,7 @@ void DivePlannerPointsModel::setupCylinders()
 			if (!cylinder_none(&(displayed_dive.cylinder[i])))
 				return;		// We have at least one cylinder
 	}
-	if (!same_string(prefs.default_cylinder, "")) {
+	if (!empty_string(prefs.default_cylinder)) {
 		fill_default_cylinder(&displayed_dive.cylinder[0]);
 	}
 	if (cylinder_none(&displayed_dive.cylinder[0])) {
@@ -175,26 +175,6 @@ bool DivePlannerPointsModel::updateMaxDepth()
 			displayed_dive.maxdepth.mm = p.depth.mm;
 	}
 	return (displayed_dive.maxdepth.mm != prevMaxDepth);
-}
-
-QStringList &DivePlannerPointsModel::getGasList()
-{
-	static QStringList list;
-	list.clear();
-	for (int i = 0; i < MAX_CYLINDERS; i++) {
-		cylinder_t *cyl = &displayed_dive.cylinder[i];
-		if (cylinder_nodata(cyl))
-			break;
-		/* Check if we have the same gasmix two or more times
-		 * If yes return more verbose string */
-		int same_gas = same_gasmix_cylinder(cyl, i, &displayed_dive, true);
-		if (same_gas == -1)
-			list.push_back(get_gas_string(cyl->gasmix));
-		else
-			list.push_back(get_gas_string(cyl->gasmix) + QString(" (%1 %2 ").arg(tr("cyl.")).arg(i + 1) +
-				cyl->type.description + ")");
-	}
-	return list;
 }
 
 void DivePlannerPointsModel::removeDeco()