+++ CHANGELOG.md
@@ -1,2 +1,4 @@
+- Desktop: Automatically sync with cloud when going online
+- Small UI fixes
+- Fix bug when using remembered BT/BLE addresses in dive computer download
 - Fix bug in cylider pressure lines
-- Add changes after 4.7.6 above this line